feat: Fixed import  to import ExploreGithub component from correct path, added Search in organizations feature, fixed stopId attribute  from 'fork' to 'explore' and added Find by URL feature
